Add files via upload
The dashboard delves into the diversity and inclusion metrics of a company with a total workforce of 500 employees, comprising 295 men and 205 women. The analysis utilizes key indicators such as average ratings, new hires, resignation rates, and promotions to assess the organization's commitment to fostering an inclusive workplace.
In a workforce of 500 employees, comprising 295 men and 205 women, average ratings reveal a slight gender disparity—men at 2.06, women at 1.89. This indicates a potential need for equal opportunity and an inclusive work environment.

Turnover rates, with 66 new hires and a 9.40% resignation rate (47 employees), prompt a closer look into retention strategies and engagement opportunities.

Encouragingly, promotions rose from 36 in FY 2020 to 51 in FY 2021. However, a nuanced analysis of gender distribution in promotions is essential for evaluating true gender equity in career advancement.
A Closer Look at Promotions, Resignations, and New Hires:
In the pursuit of workplace equity, a detailed examination of gender dynamics reveals intriguing patterns across promotions, resignations, and new hires within the organization.

Promotions:
In FY 2020, promotions showed a disparity, with 22% female and 78% male, but this shifted positively in FY 2021 to 35% female and 65% male. This progression indicates a noteworthy stride toward gender balance in career advancement.

Resignations:
The resignation landscape displays a slight skew, with 45% female resignations compared to 55% male. Understanding the reasons behind these departures is crucial for targeted retention efforts.

New Hires:
New hires present a more balanced picture, with 52% female and 48% male. This showcases an inclusive approach to talent acquisition, promoting gender diversity from the outset.

Part-Time Workers:
A closer examination of part-time workers, constituting 7% of the workforce, unveils a gender disparity. Of the part-time employees, 94% are female (33) and 6% are male (2). Addressing this gender skew in part-time roles may contribute to overall workplace equity.

In conclusion, the organization's commitment to gender diversity is evident in the positive shift in promotions and the balanced approach to new hires. However, the disparities in part-time roles and resignations signal areas for further exploration and potential improvement in fostering an inclusive and equitable workplace for all employees.